More tests, new test binaries, pipe support, started implementing registers

+++ src/lib.rs
@@ -1,7 +1,10 @@
-#[cfg(test)]
-mod tests;
+mod pipe;
+mod registers;
+
+pub use { pipe::*, registers::* };
 
 use std::ffi::CString;
+use std::process::exit;
 
 use nix::sys::{
     signal,
@@ -17,7 +20,7 @@ use libc::{
 
 pub type Result<T> = std::result::Result<T, GadbErr>;
 
-#[derive(Debug)]
+#[derive(Debug, PartialEq)]
 pub struct GadbErr {
     msg: String
 }
@@ -28,12 +31,21 @@ pub fn error<T>(msg: &str) -> Result<T> {
     })
 }
 
+pub fn error_os<T>(msg: &str) -> Result<T> {
+    Err(GadbErr {
+        msg: os_error_with_prefix(msg)
+    })
+}
+
+pub fn os_error_with_prefix(prefix: &str) -> String {
+    String::from(prefix) + &": " + &std::io::Error::last_os_error().to_string()
+}
 impl std::fmt::Display for GadbErr {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         write!(f, "{}", self.msg)
     }
 }
-#[derive(PartialEq, Clone)]
+#[derive(PartialEq, Clone, Debug)]
 pub enum ProcessState {
     Stopped,
     Running,
@@ -41,6 +53,18 @@ pub enum ProcessState {
     Terminated    
 }
 
+impl std::fmt::Display for ProcessState {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{}", match *self {
+            ProcessState::Stopped => "stopped",
+            ProcessState::Running => "running",
+            ProcessState::Exited => "exited",
+            ProcessState::Terminated => "terminated",
+        })
+    }
+}
+
+#[derive(Debug)]
 pub enum StopInfo {
     Signal(Signal),
     ExitCode(i32)
@@ -54,10 +78,13 @@ impl std::fmt::Display for StopInfo {
         }
     }
 }
+
+#[derive(Debug)]
 pub struct StopReason {
     newstate: ProcessState,
     info: StopInfo
 }
+
 impl StopReason {
     /*
     pub enum WaitStatus {
@@ -106,44 +133,75 @@ impl std::fmt::Display for StopReason {
     }
 }
 
+#[derive(Debug)]
 pub struct Process {
     pub pid: Pid,
-    autoterm: bool,
+    pub autoterm: bool,
+    pub attached: bool,
     pub state: ProcessState
 }
 
 impl Process {
     pub fn launch(cmd: &str) -> Result<Self> {
-        return Self::launch_args(cmd, Vec::new());
+        return Self::launch_args(cmd, Vec::new(), true);
     }
-    pub fn launch_args(cmd: &str, args: Vec::<String>) -> Result<Self> {
+    pub fn launch_noattach(cmd: &str) -> Result<Self> {
+        return Self::launch_args(cmd, Vec::new(), false);
+    }
+    pub fn launch_args(cmd: &str, args: Vec::<String>, attach: bool) -> Result<Self> {
         let Ok(cmd_c) = CString::new(cmd) else {
             return error("could not read cmd");
         };
+        let mut pipe = Pipe::pipe(true).unwrap();
         let pid: i32;
         unsafe {
             pid = fork();
         }
-        if pid == 0 {
-            let _ = ptrace::traceme();
-            let args_cstr: Vec<CString> = args.iter().map(|s| CString::new(s.clone()).unwrap()).collect();
-            let mut args_ptr: Vec<*const i8> = args_cstr.iter().map(|s| s.as_c_str().as_ptr()).collect();
-            args_ptr.push(std::ptr::null() as *const i8);
-            unsafe {
-                // let r = execlp(cmd_c.as_c_str().as_ptr(), cmd_c.as_c_str().as_ptr(), std::ptr::null() as *const i8);
-                let _r = execvp(cmd_c.as_c_str().as_ptr(), args_ptr.as_ptr());
-                return error(&format!("error calling exec: {}", std::io::Error::last_os_error().to_string()));
-            }
+
+        fn exit_with_error(pipe: &mut Pipe, prefix: &str) {
+            let msg = os_error_with_prefix(prefix);
+            let _ = pipe.write(msg.as_bytes().to_vec());
+            exit(-1);
         }
 
+        if pid == 0 {
+            pipe.close_read();
+            if attach && ptrace::traceme().is_err() {
+                exit_with_error(&mut pipe, &"error calling PTRACE_TRACEME");
+            }
+
+            let mut args_cstr: Vec<CString> = args.iter().map(|s| CString::new(s.clone()).unwrap()).collect();
+            args_cstr.insert(0, CString::new(cmd).unwrap());
+            let mut args_ptr: Vec<*const i8> = args_cstr.iter().map(|s| s.as_c_str().as_ptr()).collect();
+            args_ptr.push(std::ptr::null() as *const i8);
+            
+            unsafe {
+                let _ = execvp(cmd_c.as_c_str().as_ptr(), args_ptr.as_ptr());
+            }
+            exit_with_error(&mut pipe, "error calling exec");
+        }
+
+        pipe.close_write();
+        let data = pipe.read();
+        if let Ok(data) = data {
+            if data.len() > 0 {
+                let _ = wait::waitpid(Some(Pid::from_raw(pid)), None);
+                return error(&String::from_utf8_lossy(data.as_slice()));
+            }
+        }
+        pipe.close_read();
         let mut p = Self {
             pid: Pid::from_raw(pid),
             autoterm: true,
+            attached: attach,
             state: ProcessState::Running
         };
-        let _ = p.wait_on_signal();
+        if attach {
+            let _ = p.wait_on_signal();
+        }
         Ok(p)
     }
+
     pub fn attach(pid: i32) -> Result<Self> {
         if pid <= 0 {
             return error("invalid pid: 0");
@@ -156,30 +214,36 @@ impl Process {
         let mut p = Self {
             pid,
             autoterm: false,
+            attached: true,
             state: ProcessState::Running
         };
         let _ = p.wait_on_signal();
         Ok(p)
     }
-    pub fn wait_on_signal(&mut self) -> StopReason {
+
+    pub fn wait_on_signal(&mut self) -> Result<StopReason> {
         // passing None is equivalent to using 0 for the options
         let res = wait::waitpid(self.pid, None);
         let Ok(status) = res else {
-            eprintln!("could not wait on signal: {}", res.err().unwrap().desc());
-            std::process::exit(-1);
+            return error_os("could not wait on signal");
         };
         let reason = StopReason::from_wait_status(status);
         self.state = reason.newstate.clone();
-        reason
+        Ok(reason)
     }
-    pub fn resume(&mut self) {
+
+    pub fn res(&mut self) {
+        let _ = self.resume();
+    }
+
+    pub fn resume(&mut self) -> Result<()> {
         let res = ptrace::cont(self.pid, None);
         // TODO: error handling
         if res.is_err() {
-            eprintln!("could not resume");
-            std::process::exit(-1);
+            return error("could not resume");
         }
         self.state = ProcessState::Running;
+        Ok(())
     }
 }
 
@@ -188,12 +252,14 @@ impl Drop for Process {
         if self.pid == Pid::from_raw(0) {
             return;
         }
-        if self.state == ProcessState::Running {
-            let _ = signal::kill(self.pid, Signal::SIGKILL);
-            let _ = wait::waitpid(self.pid, None);
+        if self.attached {
+            if self.state == ProcessState::Running {
+                let _ = signal::kill(self.pid, Signal::SIGKILL);
+                let _ = wait::waitpid(self.pid, None);
+            }
+            let _ = ptrace::detach(self.pid, None);
+            let _ = signal::kill(self.pid, Signal::SIGCONT);
         }
-        let _ = ptrace::detach(self.pid, None);
-        let _ = signal::kill(self.pid, Signal::SIGCONT);
         if self.autoterm {
             let _ = signal::kill(self.pid, Signal::SIGKILL);
             let _ = wait::waitpid(self.pid, None);
